fix: Move load_all_tokens to simulation's utils
