+++ src/lib.rs
@@ -24,4 +24,4 @@ pub mod evm;
 pub mod models;
 pub mod protocol;
 pub mod serde_helpers;
-pub(crate) mod utils;
+pub mod utils;
